Add suggestion to use mkvmerge on error.
Often, the original source mkv files have problems.
It may not even be mkv, but simply named mkv.
When remembrall fails, one of the most common issue
is that the mkv was encoded improperly or is not an mkv file.
Running mkvmerge will test the validity of the file,
and automatically try to fix the file.
Once mkvmerge generates a new file,
rememberall can be run on the new file as normal.